pyproject.toml
[tool.poetry]
name = "rstblog-content"
version = "0.1.0"
description = ""
authors = ["Kevin Cuzner <kevincuzner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^3.4"

[build-system]
reqires = ["poetry>=1.0"]
build.backend = "poetry.masonry.api"
